New API for signing avoids to manipulate a public key.
Changes:
 - The PrivateKey struct now contains the private key (seed) and
   a copy of the public key.
 - The PrivateKey struct do not allow to manipulate the public key.
 - The PrivateKey implements crypto.Signer, if someone needs it.

Fixes #81.
