+++ sign/doc.go
@@ -1,0 +1,2 @@
+// Package sign provides variety of digital signature schemes.
+package sign
